ai_service/requirements.txt
flask
pandas
scikit-learn
mysql-connector-python
flask-cors
python-dotenv